Ignore key order for APC bespokes
Summary:
We've found that a lot of times the data structures we store in APC have the same keys but in different orders.

That causes bespoke escalation, because we don't match the new order against the profile for the old one. Use a commutative APC key to avoid this issue.

Reviewed By: colavitam

Differential Revision: D28238762

fbshipit-source-id: 72addd1b0d41a0795973dda04b1d27b8361101a3