Load unit paths lazily in RepoFile
Summary:
Right now we keep a mapping of unit paths to the metadata needed to
load the unit. These paths are static strings. Unfortunately this can
cause the static string table to grow too large. Instead load the unit
paths lazily. The key idea here is that instead of storing a mapping
of the paths to metadata, we store a mapping of path SHA1 to
metadata. The actual path is stored in the metadata and that's only
loaded when actually needed. Before the path is loaded, we store the
offset where the path can be found. Once loaded, it's the actual
StringData* for the path. Using SHA1 in place of the path is almost
equivalent because of the chances of collision are almost
nill. Nonetheless, we do a sanity check after we load the unit to make
sure the paths actually match.

Reviewed By: oulgen

Differential Revision: D28249302

fbshipit-source-id: d39b5c8f4fab7fc3e48d3e6a10234bc15a8c630d