Profile all object constructors
Summary:
It turns out that almost everything is working for these constructors - we even make these props bespoke! - except that we failed to profile all object constructor bytecodes, leading us to incorrectly turn some sinks into vanilla sinks.

Profiling an object is a simple function on an Object* alone, so it's easy to apply everywhere it's needed.

Reviewed By: colavitam

Differential Revision: D28257384

fbshipit-source-id: 88796e4cbfd52017a0e9d3b48482eea66fd3bc16