APC backoff for unprofiled sources
Summary: Post-layout-selection, if we don't have a selected layout for an APC array, preserve its incoming layout. This approach will work well when we store many closely-related struct arrays into APC; we'll make the right decision a) if we saw an exact match on struct fields at profiling time OR b) if the constructor for the initial, counted array was converted to a struct.

Reviewed By: colavitam

Differential Revision: D28263319

fbshipit-source-id: f84eeedd496bc8b98e354618a8292928c224ec0d