Rename isAbstract() to isAbstractAndUninit()
Summary:
This diff is a no-op, but it exposes all of the former callsites of `PreClass::Const::isAbstract()` and `Class::Const::isAbstract()`. These currently look at the `isAbstract` bit, but the bit is only set when an abstract type constant does not have a default.

Follow up diffs will decouple the bit and the presence of default.
- The old behavior will gain a secondary assertion, namely `isAbstract() && !val().is_init()` (see {D28210469}).
- A new `isAbstract` function will be defined that exclusively checks the bit.
- I will gradually migrate the callsites of `isAbstractAndUninit` to `isAbstract` using runtime options.
- `isAbstractAndUninit` will be removed; places that truly need to check the presence of a default will do so explicitly, though they should be few.

Reviewed By: oulgen

Differential Revision: D28210195

fbshipit-source-id: 704c04474a1080c6155b05d2db5ea531a870c540