Housekeeping: remove a few more vestiges of APC prime
Summary:
No change in functionality.

Renamed ConcurrentTableSharedStore::bumpTTL() to extendTTL() because
it implements the `apc_extend_ttl()` function and is not directly related to the
newer `$bump_ttl` parameter to `apc_add()` and `apc_store()`.

Renamed `adjust_ttl()` to `apply_ttl_limit()` and removed the `overwritePrime`
parameter, which was only used by one caller.

Fixed a few doc comments.

Reviewed By: binliu19

Differential Revision: D28143934

fbshipit-source-id: 1abc5f4253679f7f19f5d65e5aaf944db9bf7560