Use abstract keyword to determine abstractness for type constants
Summary:
Following up on {D28210195 (https://github.com/facebook/hhvm/commit/f7fc7e24949296e4a15ec776419dd295c45c374e)}, this introduces a new `isAbstract()` function that just returns if the bit is set.

Added condition that the default is missing `isAbstractAndUninit`.

Updated the lone use site of the isAbstract bool in hhbbc/index.cpp to also be `&& !cns.val`. Otherwise, an abstract type constant with a default would now be dropped if it got inherited.

Added the first use site of `isAbstract` in the disassembler, which now emits isAbstract and the default for abstract type and context constants with defaults.

Reviewed By: oulgen

Differential Revision: D28044680

fbshipit-source-id: f85292fefb42bc56248feb7c2ea8e4097b2768b8