Don't make LoggingArrays for null RuntimeStructs
Summary:
We can wind up producing a null RuntimeStruct when a registered key may undergo
an intish cast. Currently, this will wind up creating a LoggingProfile
associated with a LocationType::Runtime and a null RuntimeStruct. Instead,
don't emit a LoggingArray.

As a side note, it's not clear to me whether the better solution for sources
like this is to make them unsampled or to assign them a special LocationType
that can't be specialized by layout selection. We can revisit this.

Differential Revision: D28276589

fbshipit-source-id: 14c37b0dbb646cd66519740626a453d305ba2577