Add builtins to collect a list of functions called during a sandbox request
Summary:
Add a new feature for sandbox mode (which when Eval.EnableFuncCoverage=1) allows a set of all called functions between the time that `enable_function_coverage()` and `collect_function_coverage()` (the latter returning a dict mapping said function names to the files in which they were defined) are called.

Collection does not require the interpreter and uses an RDS local flag in the JIT to avoid calling into the runtime when coverage has already been collected.

Reviewed By: ricklavoie, jano

Differential Revision: D27921403

fbshipit-source-id: eada7fa6a6248096676385ab6d9b364027bcd96a