log Server_hung_up
Summary:
The problem the previous diffs have been trying to solve is that our logging of Server_hung_up was scattered all over the place. Those diffs have gathered it into a central place.

Now with this diff we're in a position to log it properly. This diff merely adds logging.

Differential Revision: D28139927

fbshipit-source-id: f5edf86170e6405a3c428452bd3fcc9c139ad184