refactor ClientConnect.with_server_hung_up
Summary:
Our logging of "monitor connection failure" is a bit rough because it gets lumped with all other kinds of server exceptions into just Exit_status.server_hung_up_should_retry for our logs.

In this diff I'm trying to disentangle the code a little to make it more plain what happens.

All I've done is moved the "Printf.eprintf" statement up a little, from the place that catches the exception to the place that generates it. And renamed a function

Differential Revision: D28139803

fbshipit-source-id: ac8517330f461f607ae1a823e8a3b50086e3dec6