Distinguish "monitor can't handoff to server" from other cases of server-hung-up
Summary:
1. When the monitor tries to handoff to the server it does a "select 0.5" call on its end of the monitor->server pipe, waiting up to 0.5 seconds until that pipe is ready to receive something.
2. If it isn't ready (which I suspect happens because the pipe is full) then the monitor will try again, up to 8 times, before giving up. It gives up by closing its end of the client->monitor/server pipe.
3. The client sees this as EOF. It doesn't see a server-finale file. It therefore believes that the server has crashed badly. It exits with exit code Server_hung_up_should_retry (6).
4. find_hh.sh interprets exit code 6 to mean it should retry.

All of that seems fine. I'm just wondering if I can make it more explicit what's happening, so I can understand how common this path is. In particular, I'd like for the client to know the difference between "monitor failed to handoff" and "monitor succeeded to handoff but then server hung up".

This diff changes it so that, if the monitor finds the server too busy for handoff, then it communicates this fact back to the client with a Monitor_failed_to_handoff message. We log this. It still gets translated into the same exit code 6 for benefit of external callers.

Note: the client doesn't know whether it's going to get on its FD a Hello message from the server (after a successful handoff), or a Monitor_failed_to_handoff message from the monitor (because of server being busy). The two messages must therefore be part of the same ADT. That's why I've added Monitor_failed_to_handoff to ServerCommandTypes.message_types

Differential Revision: D28135756

fbshipit-source-id: 6e6dfadfc33d44847be061e53bc9f451b91d5ba5