track whether an IDE was hooked up to get push diagnostics
Summary:
I have a vision that we should monitor how long hack takes to run.

For "hh" commands that's easy. We already track those with CLIENT_CHECK_START and CLIENT_CHECK, and we record in this event telemetry about the most recent typecheck to have completed. The --from argument distinguishes command-line users from scripts.

But another class of waiting is when you're primarily an IDE user, and you save a file, and that kicks off a typecheck. It's not actually even clear whether users are waiting in this case. Some users just save out of reflex and don't care to wait until all errors are done. Others sit there waiting for the spinner to finish so they can see all errors. Others, even though they're in the IDE, switch to the command-line so they can see typechecking progress and know when it's done.

How can we track those IDE waitings?

This diff provides a solution. When a whole-program typecheck happens, it pushes diagnostics to a  connected IDE client, if present. With this diff, we include information about that push in the normal typecheck telemetry. Therefore:
1. If a user saves a file which kicks off a typecheck, the RECHECK_END event `push_diagnostics` field now indicates whether an IDE was connected (and therefore presumably had a progress spinner for the duration of the check). This will be our proxy for how long user spend in the IDE waiting for a typecheck. (we don't know if they're actively waiting for the results, or just don't care, but we can never know that).
2. If a user types "hh" at the command-line, the CLIENT_CHECK event includes telemetry about the most recent typecheck, and thus indicates by the same `push_diagnostics` field whether an IDE had shown a spinner prior to or during the user invoking hh at the command-line.

Differential Revision: D28294800

fbshipit-source-id: 1c0d3266ca40324e42fe15fb4f358d92364602c2