manually implement DeserializeInArena for BStr
Summary:
`BStr` is a wrapper of [u8], the default deserialization implementation only works for some data format, for example, it works FluxBuffer, but it doesn't work for JSON.

## Why can't we use default deserialize?

The following is the implementation, the lifetime `'a` is bounded by lifetime time `'de`, this means that `&'a BStr`(or `&'a [u8]`) is actually referencing content from the underlying data. This can explain why deserialization works for some data format. `&'a BStr` is just a sequence of bytes, if it references the underlying data, the the data must be binary format, e.g. FluxBuffer. It *can't* references data from JSON, since sequence of bytes is represented as "[97,115,100,102]" literally. I actually tried it, it panics at runtime with error, "invalid type: sequence, expected a borrowed byte string", this is because it doesn't implement `visit_seq`.

```lang=rust
    impl<'a, 'de: 'a> Deserialize<'de> for &'a BStr {
        #[inline]
        fn deserialize<D>(deserializer: D) -> Result<&'a BStr, D::Error>
        where
            D: Deserializer<'de>,
        {
            struct BStrVisitor;

            impl<'de> Visitor<'de> for BStrVisitor {
                type Value = &'de BStr;

                fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
                    f.write_str("a borrowed byte string")
                }

                #[inline]
                fn visit_borrowed_bytes<E: Error>(
                    self,
                    value: &'de [u8],
                ) -> Result<&'de BStr, E> {
                    Ok(BStr::new(value))
                }

                #[inline]
                fn visit_borrowed_str<E: Error>(
                    self,
                    value: &'de str,
                ) -> Result<&'de BStr, E> {
                    Ok(BStr::new(value))
                }
            }

            deserializer.deserialize_bytes(BStrVisitor)
        }
    }
```

## How can we implement `DeserializeInArena` for `&'a BStr`?

Since `BString` *does* implement `Deserialize`, a naive implementation is to leverage `BString`.

```lang=rust
impl<'arena> DeserializeInArena<'arena> for &'arena bstr::BStr {
    fn deserialize_in_arena<D>(arena: &'arena Bump, deserializer: D) -> Result<Self, D::Error>
    where
        D: Deserializer<'arena>,
    {
        let v = <bstr::BString as Deserialize>::deserialize(deserializer)?;
        let bytes: &[u8] = arena.alloc_slice_copy(v.as_ref());
        Ok(bytes.into())
    }
}
```
This implementation looks not bad and concise, ship it? NO. This implementation is not optimal in terms of performance, it always allocates on heap in order to create a `BString`, then converts to a slice and copies it to arena. For JSON format, this is fine. But for binary data format,  the intermediate `BString` is wasted, because arena only needs a slice of bytes. A solution is to implement `visit_borrowed_bytes` and `visit_borrowed_str` and also implements `visit_seq`, which is very similar to the one for `BString`.

Reviewed By: dtolnay

Differential Revision: D28183183

fbshipit-source-id: 2bf2696a116259fac9df3655e682a47961995675