Add decl serialize/deserialize round trip tests
Summary:
This diff added a test and a rust binary,

the rust binary is `hphp/hack/test/decl/serialize`, it takes a dir and run ser/de round trip on decls of each php file. It reports error if round trip fails or it reports benchmark on all ser/de providers. The benchmark includes serialized output size in bytes, and ser/de time.

this diff also implements `DeserializeInArena` for more types in order to implement `DeserializeInArana` for `Decls`.

Note: flexbuffers ser/de failed in this diff, and I commented out it, it will be fixed in the next diff. D28290660

Reviewed By: shayne-fletcher

Differential Revision: D28220648

fbshipit-source-id: 35318529961ace28c967f7949d0cc80230474513