Spurious error when accessing method on class with class-level where constraints
Summary:
When `enable-class-level-where-clauses` is set, we erroneously reject calls to methods that exist on an interface when that interface also specifies `where this as ...`. The problem is the incorrect use of `Errors.try_with_error` (which accumulates errors) instead of `Errors.try_` (which is more like try...catch).

Error messages are less than desirable (it would be nice to be told that Hack looked in `C` when an interface specifies `where this as C` rather than just be told that the method is missing from the interface) but this is already an issue with intersection types, for example, and should be fixed more generally.

Reviewed By: madgen

Differential Revision: D28183801

fbshipit-source-id: 412502cafa15c9dc87c094237f825b15af3ff9d4