T66877858 is not a typehole, move the test
Summary:
This test was expected to highlight a typehole in the handling of the dynamic type, but actually Hack is not inferring Ref<int> as type of `$r`.  It is instead inferring a more permissive type.

Looking again at this task with enable-sound-dynamic-type enabled, the
test still returns `no error`, and the reason is the following:

- `new Ref(42)` in (1.) introduces a new type variable for the content of Ref, with `int` as lower bound;

- (2.) adds an additional `dynamic` lower bound

- `$r->get()` returns the type variable with `int | dynamic` as lower bound and no upper bound

- since `int` is enforceable, sound-dynamic relaxes the type of the
  argument of expect_int to `int | dynamic`.  The subtype check to call
  expect_int succeeds.

If `$r` is initialised with `new Ref<int>(42)`, then Hack returns an error
as expected (both with and without sound dynamic).

So, even with sound dynamic, this is a surprising behaviour, but technically not typehole nor a bug.

Differential Revision: D28322824

fbshipit-source-id: 7486c03eece2f84c95b764d34244b7737e8fcbfc