Use diamond codegen at low-sample-count sinks
Summary:
In this diff, we introduce a mechanism to JIT a diamond at a bespoke sink rather than an exit. To make the decision, we would like the sink to pass all of three heuristics:
1. We should see no more than SideExitMaxSources sources
1. For each source, we should see at least SideExitMinSampleCount samples
1. The probability of the selected layout should pass a high SideExitThreshold

We implement items 1 and 2, but leave item 3 (which is more complicated) to a follow-up diff.

Reviewed By: colavitam

Differential Revision: D28309009

fbshipit-source-id: 031940d6231cd743bde59ed30a94363e27849e08