Add the option to remove positions while printing ASTs
Summary:
Positions in AST printing sometimes gets in the way of debugging and testing. This diff provides a way to turn it off.

Where possible this diff strips the position altogether and otherwise it replaces it with `Pos.none`.

Positions are eliminated with `--no-print-position` flag (combined with `--tast` or `--nast`).

Reviewed By: CatherineGasnier

Differential Revision: D28257280

fbshipit-source-id: 2e51c0c3eef6e843fad180770bb901fcdf71ba3c