Rewrite comment about too many requests
Summary:
There's a lengthy comment in MonitorConnection.ml about "DDOS attacks" where the monitor can't cope with the incoming rate of requests. The comment imagines that an arbitrarily high number of messages can be sent to monitor and server and the only limiting factor is the rate at which the monitor processes them.

But really, the problem is a different one -- backpressure. The server doesn't store its own arbitrarily-growing queue. Nor does the monitor. Therefore, some messages we post to those things are going to experience a full queue. Experimentally the queue size for monitor->server handoff seems to be between 20 and 40 messages.

I've rewritten the comment.

Differential Revision: D28157423

fbshipit-source-id: 001617e1dcbd62a3e1b8f958dfc03e614fb8e3e1