monitor needn't retry handoff; just one attempt with longer timeout is good
Summary:
How does the monitor handoff a message to the server? The way it was written back in Oct 2015 D2583617 (https://github.com/facebook/hhvm/commit/d9a4163ff7f456ac08b5dd10ce489414aee9f94a) is that (1) it attempts to send the message with a timeout of 0.5s, (2) it retries up to 8 times if that times out, or gets an exception.

Retrying 8 times with a timeout of 0.5s is weird. Better just to try a single time with a timeout of 4s.

Also I don't see any reason to retry if we got an exception.

Differential Revision: D28225137

fbshipit-source-id: f4256c89c98f8d8c30b0cee513cc4e633f8319b6