Add HH\Readonly\as_mut as a special builtin to HackC
Summary:
This diff adds a special builtin function HH\\Readonly\\as_mut, which converts a value type like int to mutable from readonly.

Only readonly integers and other value types should be able to be passed into this function, but we haven't built the support for that yet. That will come in a diff on top of that for the typechecker.

Reviewed By: oulgen

Differential Revision: D28073468

fbshipit-source-id: 00995cb1086f62e0c1169b45118387ee475f108a