Reserve 14 bits of StringData for color information
Summary:
Currently, the only use of m_aux16 for static StringDatas is to store a flag
indicating whether the string is a symbol or not. We intend to use the
lower 14 bits to store the color of the string. A subsequent diff will
elaborate on the purpose of the coloring, but it will be used to implement
perfect hashing for StructDicts.

Move the symbol flag to the MSB of m_aux16, and reserve the lower 14 for this
color.

Differential Revision: D28330347

fbshipit-source-id: 0cc5568dccb6a0b50edfc236bdf006431083a157