Key order pruning with array count awareness
Summary:
One big chunk of key order pruning improvements.

The previous strategy for pruning keys avoided removing keys that would
invalidate a large number of recorded key orders. However, this would not
necessarily prevent escalating: if we always append 100 keys to a dictionary,
then the last 4 keys can be pruned without crossing the 95% threshold on
recorded key orders, but all arrays will escalate.

Instead, we use the recorded SetStr and ConstructStr events to produce an upper
bound on the number of logged arrays containing each key. When we prune keys,
we do not allow the upper bound on the number of logged arrays to cross this
threshold either.

Differential Revision: D28276439

fbshipit-source-id: d9d8607a8829aba478c4bf1c30075b2085238198