Add probability cutoff for diamond codegen
Summary: In this diff, we add logic for heuristic 3 from D28309009 (https://github.com/facebook/hhvm/commit/46c2506640b487158d12383ee17a94c50e35411e) to layout selection. The tricky part here is that we have to track the probability of each decision. As we do so, we also deal with monoarrays: we union-bound probability computations for them, rather than just assuming that mistakes are correlated.

Reviewed By: colavitam

Differential Revision: D28308585

fbshipit-source-id: c5f8ec2bc894cab735b64624051a088a79cf93e7