Greedy coloring for perfect hashing
Summary:
This diff implements the first part of perfect hashing: coloring the strings of
our selected layouts. Our perfect hashing implementation will use the color of
a static string as its hash key. In a given layout, each key must have a
distinct color to avoid hash collisions.

To color the strings, we create an interference graph where nodes are keys and
edges are present between any two keys that appear in the same layout. We then
proceed with a greedy coloring, selecting the node of highest degree and
assigning it the smallest color not belonging to one of its adjacent nodes.

We currently cap the number of colors used at 255. This means that not every
set of layouts will be colorable. To address this issue, we binary search on
the prefixes of our layouts, ordered by weight, to find the longest prefix that
is colorable within our limit. Layouts after this prefix will be removed. Some
quick experiments have shown that 255 is sufficient, but we will want to
validate this in production and adjust our cutoffs accordingly.

Differential Revision: D28330621

fbshipit-source-id: 071e2d32a606d52e2a72b407afc88d85bcb2c9c8