Remove references to preg_last_error in tests
Summary:
As announced in the HHVM 4.105 release notes, `preg_last_error` is going to be deleted soon.

HHVM tests currently use this, so references to `preg_last_error` will be removed an instead we call the `_with_error` equivalent for preg functions. Some test cases become obsolete and are removed.

Reviewed By: oulgen

Differential Revision: D28231470

fbshipit-source-id: aed9e8d3e606d49c6615bc70cc8267b2d4e8ac73