Fix stream_copy_to_stream truncation at INT_MAX
Summary:
Rather than treating INT_MAX as the copy limit, we can run till we find an EOF if `maxlength` is set to -1.

Because we can't actually read the full string in this case (strings > 2 GB aren't supported in HHVM), we can test it by writing a big file, copying it, and verifying the part past the 2GB boundary was copied.

Reviewed By: ottoni

Differential Revision: D28313569

fbshipit-source-id: 11f12f3fa44ffccc058c50acae968fccf985a8fd