fix EC keys to export as PKCS#8 instead of PKCS#1
Summary:
We live in a PKCS#8 age.  For what I assume are historical tech reasons (if RSA else EC), this extension is special-casing EC key export to be PKCS#1.

Put another way, it looks like this:

```
# ONLY when using an EC key
-----BEGIN EC PRIVATE KEY-----
...
-----END EC PRIVATE KEY-----
# literally any other key type
-----BEGIN PRIVATE KEY-----
...
-----END PRIVATE KEY-----
```

This diff resolves that by changing the OpenSSL function call from a key-specific one (results in PKCS#1), to just use the preexisting EVP generic key export which is always PKCS#8.

Reviewed By: ottoni

Differential Revision: D28034769

fbshipit-source-id: 56f57023fe0526df2c0b378d939a29456df1f713