Fix widening_union
Summary:
widening_union is a version of union_of used when we've triggered our
infinite cycling heuristics. It performs an union, but ensures that we
eventually reach a fixed point. In practice it does this by
restricting how deeply a type can nest. As an optimization, it will
just return one of the inputs if either is a subtype of the
either. However this is actually incorrect, as this can result in
chains of infinitely growing types. Remove this optimization.

Reviewed By: ottoni

Differential Revision: D28376411

fbshipit-source-id: e9a4315e3b1a63c532a005bb2ec14ab1dd7ff48c