Remove ability to copy slabs for reading at an arbitrary base address
Summary: Now that we mount the decl-service cache at a fixed address (both in the decl service and in workers), there is no need for us to rebase cache items with respect to the address at which the prototype mounted the cache. We can just rebase them to their current address instead, and trust that they'll have the same address in the prototype process. This functionality (in `ocamlrep::slab::copy_slab`) was added specifically for this use case, and doesn't seem very generally useful, so this diff just removes it altogether.

Reviewed By: bobrenjc93

Differential Revision: D28156972

fbshipit-source-id: 0b97ca3b9f57503a97042b3c89f6d9433ea81e58