Set directly into Sets instead of using lvals during Unserialization
Summary:
Currently we get an lval into our MixedArray with `findForUnserialize()` and then we `tvSet` the typedvalue of that key into our set. Since the value is just the key, we can just directly set into it without needing an lval.

Note that for sets we do need to do anything on conflict because the element is the same as the key. Therefore, we can specialize this case to be an noop on conflict.

This eliminates another spot where we would need to record the datatype we are putting into our mixed array.

Follow-up to the comment on D28209743

Reviewed By: ricklavoie

Differential Revision: D28389776

fbshipit-source-id: 2e907c8cabb4cf9ac7d10377567c8de0908b4d64