fix printing NASTs for multifile test files
Summary:
Previously, `hhst --nast` on a file with multiple contents (i.e. with `//// file1.php` comments) would raise an exception at Typing l.1343 saying the filename was not found in the map of NASTs. The logging that this diff adds would show:
```
Could not find nast for file myfile.php.
Available nasts:
  myfile--file1.php
  myfile--file2.php
```
So the proper list of files to pass is not `filenames` but the keys of `file_contents` which results from splitting the input file.

Reviewed By: madgen

Differential Revision: D28379744

fbshipit-source-id: 9a7790d43e5979e65ce4df7a9ec2c64158cd3063