Use a priority-based traversal in HotCFG region former
Summary:
This diff switches the HotCFG region formation algorithm to use a priority-based traversal of the TransCFG instead of a depth-first search (DFS). The DFS approach can lead to pathological behavior when the TransCFG is too large and the region is truncated due to the bytecode-size limit. The problem is that, if you go too deep on a path in the CFG without traversing many of the predecessors that can reach the blocks in that path, and then truncate the region before getting a chance to visit those predecessors, regionizeFunc will then have to create other large region that duplicate that long, deep path because one region cannot jump into the middle of another.

This diff addresses that problem by inserting blocks in the region giving priority to blocks that already had all their predecessors visited. To avoid issues with loops, backedge sources are not included in the set or predecessors that needs to be visited prior to visiting a block. Among all the blocks that are eligible to be visited (i.e. considered for inclusion in the region), we give priority to blocks with higher profile counts. One other potential issue with this approach is that blocks may no be included in the region if they have very low weights and, yet, we don't want them to prevent their successors from being included in the region. This is addressed by visiting even blocks that will not be inserted in the region.

Reviewed By: paulbiss

Differential Revision: D28401526

fbshipit-source-id: 6fd833b89d069684670ca686f776b0d3060ed190