Fix attempts at passing expressions as atoms
Summary:
In the current implementation, one could succeed to pass a normal constant
as an atom if it was in a local, or returned by a function call.
This diff prevents this, so `<<__Atom>>HH\MemberOf` parameter can only
accept literal labels, like `#Foo`.

If one wants to pass an atom around, it must use the `HH\Label` type, cause
using `__Atom` triggers an automatic resolution to the constant, and the
actual label is not exposed.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D28118728

fbshipit-source-id: 31047c56b573fedcdd1c254821a2d14c0d615570