Density can never be too low on entry to makeRoom()
Summary:
A cleanup I found while investigating another optimization.

`makeRoom()` is called when a hash-collection is full and we want to add an element to it. If the density is too low, we compact, otherwise we grow. However, we always will want to grow in this case. Density too low would mean half our elements are tombstones, and tombstones are added when we call `erase()` and `erase()` checks if density is too low on exit, and compacts if it is.

Also
-  `canMutateBuffer()` => `m_immCopy.isNull()`
- `isFull()` == `posLimit() == cap()`

Reviewed By: ricklavoie

Differential Revision: D28454595

fbshipit-source-id: ea7ce79123ac7d677233506ad0bbe01c21f2e45b