Separate top-level functions from lambdas in the AAST
Summary:
Previously, an `Aast.fun_` would represent both top-level function as well as a lambda. This was confusing, as several of the `fun_` fields did not make sense for lambdas.

Instead, define a separate `fun_def` type that stores fields that only exist on top-level functions, such as `mode` and `file_attributes`.

Differential Revision: D25987602

fbshipit-source-id: 36f4d83a04871bd76d746b78bc5a4e04912fcf9e