report nested TAST `Hole`s correctly when calling `--type-error-at-pos`
Summary:
This diff updates the behaviour of `--type-error-at-pos` to report a sub typing / coercion error, represented by as TAST `Hole`, only when the position provided exactly matches the start position of the `Hole`. Previously, the command followed the same logic as `--type-at-pos` which finds the smallest enclosing type; however we do not want to find the `Hole` _containing_ a position, rather we want to report a `Hole` when it occurs at that _exactly_ subexpressions position.

This prevents erroneous errors being reported for subexpressions when there is a `Hole` surrounding their parent expressions.

Reviewed By: madgen

Differential Revision: D27998741

fbshipit-source-id: b804b1000cfaacf029dfdbb832c0b668fb478866