do not add a Hole with well-typed destructure assignments
Summary:
Fixes a bug in the assignment to a `list(...)` causing a `Hole` to be added with identical source and destination types.

If we can successfully destructure the rhs of the assignment to the lhs variable, the code currently checks that the list of optional errors is empty or else adds a `Hole`. In fact this list of optional errors will never be empty and should instead be checking that all optional errors are `None`

Reviewed By: madgen

Differential Revision: D28220678

fbshipit-source-id: 764644aa8ce7391e06bae16f27f05a565b8a9ae2