add test case for expression dependent type
Summary: When the argument in a function call is an local var, coercion is applied to a constructed expression dependent type rather than the original expression type. ~~For the purposes of `Hole`s, we want annotate with that original type.~~ It looks like we do want the expression dependent type; the fix is on the `HackAst` side of the codemod. Added a test in this diff which illustrates a case where it is necessary

Reviewed By: madgen

Differential Revision: D28254378

fbshipit-source-id: 2bb797cd9797712b7bfa3886d593970cf2386450