Fix assignment to a Map with bad key
Summary:
It is not safe to use a `dynamic` key on a `Map` whose key type is either `string` or `int`, because hhvm only enforces that the key is an `arraykey`. For example,

```
function f(dynamic $d) : Map<string, int> {
  $output = Map<string, int> {};
  $output[$d] = 2;
  return $output;
}

...

  $m = f(1);
  $k = $m->keys()[0]; // k : string but contains an int
```

This adds a check that the `Map` has key type of exactly `arraykey`.

Reviewed By: madgen

Differential Revision: D28324128

fbshipit-source-id: a6abbdfa4e1f917cd20cc64bc6315b350abbe933