Colocate RuntimeStruct and its fields
Summary:
RuntimeStruct is a hot data structure, and the fields we access most often are the slots. We colocate these slots with the array.

We keep the keys and the number of fields, which we only use in debug mode, in a separate CompactVector allocation. (You can trace and see that m_fields and getKey() are only read on creation, deserialization, and in assertions.)

Reviewed By: colavitam

Differential Revision: D28491274

fbshipit-source-id: 1fb16c1046216fca57a63d8e2e382334379bb61a