Split between where constraints
Summary:
Give the formatter permission to add line breaks between each where constraint in a where clause.

It might be nice to put the function body's opening curly brace on a new line whenever we add line breaks between where constraints (as rustfmt does):

```
  public static function g___________________<
    <<__Enforceable>> reify Tx as A___________________,
    Ty as B___________________,
    Tz as C___________________,
  >(Ty $y, Tz $z): void
  where
    Tx::Tt = Ty,
    Tx::Tu = Tz,
    Tx::Tv as D____________________,
    Tx::Tw as E_________________________,
  {
    do_something_else($y, $z);
  }
```

This would preserve blame for the last where constraint in the list when a new constraint is added. It would also help visually distinguish the where constraints from the body of the function. I think this would be a more involved change, though.

Reviewed By: ravibhoraskar

Differential Revision: D28540670

fbshipit-source-id: 57106430102619866b6d9b047f46373079d1226f